requirements.txt
# Development dependencies
gradio
farm-haystack
boto3

# Deployment dependency
aws-cdk-lib